refactor(cli): flatten browser command module
Move the Browser command slice to the neutral command root and update its internal export and test-only fixtures without changing Browser behavior.
